Refactor program logic
Now that some of the client code is developed, it makes sense to
move it around to make the structure more logical.
