+++ vcl_client/cfg.py
@@ -7,7 +7,6 @@ import keyring
 
 CONFIG_FILE_PATH = os.path.expanduser('~/.vcl.conf')
 CONF_SECTION = 'vcl'
-
 DEFAULT_ENDPOINT = 'https://vcl.ncsu.edu/scheduling/index.php?mode=xmlrpccall'
 ENDPOINT_KEY = 'xmlrpc_base_url'
 IMAGE_LIST_KEY = 'image_list'
@@ -17,17 +16,23 @@ CONF = ConfigParser.ConfigParser()
 CONF.read(CONFIG_FILE_PATH)
 
 
-def initialize_config():
-    """Sets the initial config file options."""
+def vcl_conf(username, password, endpoint):
+    """Sets the initial VCL config options."""
     if not CONF.has_section(CONF_SECTION):
         CONF.add_section(CONF_SECTION)
+
+    set_conf(USERNAME_KEY, username)
+    set_password(password)
+    set_conf(ENDPOINT_KEY, endpoint)
+
+    write_conf()
 
 
 def get_conf(key):
     """Gets a key from the config file"""
     try:
         return CONF.get(CONF_SECTION, key)
-    except ConfigParser.NoOptionError:
+    except (ConfigParser.NoOptionError, ConfigParser.NoSectionError):
         return None
 
 
@@ -54,4 +59,10 @@ def get_password():
         return None
 
 
-initialize_config()
+def set_password(password):
+    """Stores the password in the keyring."""
+    username = get_conf(USERNAME_KEY)
+    if not username:
+        raise ValueError('Cannot set password without a username specified.')
+
+    keyring.set_password('system', username, password)
